Refactoring functions into pure function
