+++ src/index.js
@@ -1,8 +1,4 @@
 import './style.css';
-import addTodo from './modules/addTodo.js';
 import displayTodos from './modules/displayTodos.js';
-import removeTodo from './modules/removeTodo.js';
 
 displayTodos();
-addTodo();
-removeTodo();
