+++ src/modules/addTodo.js
@@ -1,25 +1,5 @@
-import displayTodos from './displayTodos.js';
 import { todoCollection } from './TodoCollection.js';
 
-const addTodo = () => {
-  const formContainer = document.querySelector('#add-new-form-container');
-  formContainer.innerHTML = `
-    <form class="add-new-task" action="">
-      <input type="text" id="todo" name="description" placeholder="Add to your list..." />
-      <button id="add-todo" type="button">
-        <i class="fa-solid fa-arrow-left"></i>
-      </button>
-  </form>
-  `;
-
-  const form = formContainer.querySelector('form');
-  form.addEventListener('submit', (event) => {
-    event.preventDefault();
-    const description = form.description.value;
-    todoCollection.addTodo(description);
-    form.description.value = '';
-    displayTodos();
-  });
-};
+const addTodo = (description) => todoCollection.addTodo(description);
 
 export default addTodo;
