+++ src/modules/displayTodos.js
@@ -1,3 +1,4 @@
+import removeTodo from './removeTodo.js';
 import { todoCollection } from './TodoCollection.js';
 
 const displayTodos = () => {
@@ -27,6 +28,31 @@ const displayTodos = () => {
     description.addEventListener('change', () => {
       todoCollection.updateDescription(index + 1, description.value);
     });
+  });
+
+  const formContainer = document.querySelector('#add-new-form-container');
+  formContainer.innerHTML = `
+    <form class="add-new-task" action="">
+      <input type="text" id="todo" name="description" placeholder="Add to your list..." />
+      <button id="add-todo" type="button">
+        <i class="fa-solid fa-arrow-left"></i>
+      </button>
+  </form>
+  `;
+
+  const form = formContainer.querySelector('form');
+  form.addEventListener('submit', (event) => {
+    event.preventDefault();
+    const description = form.description.value;
+    todoCollection.addTodo(description);
+    form.description.value = '';
+    displayTodos();
+  });
+
+  const clearButton = document.querySelector('.clear-completed');
+  clearButton.addEventListener('click', () => {
+    removeTodo();
+    displayTodos();
   });
 };
