+++ src/modules/removeTodo.js
@@ -1,13 +1,5 @@
-import displayTodos from './displayTodos.js';
 import { todoCollection } from './TodoCollection.js';
 
-const removeTodo = () => {
-  const clearButton = document.querySelector('.clear-completed');
-
-  clearButton.addEventListener('click', () => {
-    todoCollection.clearCompleted();
-    displayTodos();
-  });
-};
+const removeTodo = () => todoCollection.clearCompleted();
 
 export default removeTodo;
